Replace MD5 filename with timestamp
The intention of using a MD5 as the filename was to speed up the
creation of .gitignore files by checking if a file already existed.
However, several flaws in the implementation and logic make this
approach impractical.

First, since the names were not ordered, multiple files with the same
content could still be created by changing the order of the arguments.

Second, no mechanisms was implemented to invalidate existing files, e.g.
when the user updated the source templates and wanted to create a new,
updated .gitignore file.

Both problems could be easily solved to maintain the same approach, but
then again, speed is not yet a concern, and for now it is much easier to
just use timestamps instead of fixing both bugs.